requirements.txt
scapy-python3
netifaces
requests
maxminddb-geolite2
